+++ Sources/Views/Cells/MessageCollectionViewCell.swift
@@ -103,7 +103,7 @@ open class MessageCollectionViewCell: UICollectionViewCell, CollectionViewReusab
         let messageColor = displayDelegate.backgroundColor(for: message, at: indexPath, in: messagesCollectionView)
         let messageStyle = displayDelegate.messageStyle(for: message, at: indexPath, in: messagesCollectionView)
         
-        displayDelegate.configureAvatarView(avatarView: avatarView, for: message, at: indexPath, in: messagesCollectionView)
+        displayDelegate.configureAvatarView(avatarView, for: message, at: indexPath, in: messagesCollectionView)
 
         messageContainerView.backgroundColor = messageColor
         messageContainerView.style = messageStyle